+++ src/utils/global-source-files.js
@@ -4,7 +4,7 @@ const path = require('path');
 const processCwd = process.cwd();
 
 module.exports = [
-  `!${path.join(processCwd, '**/node_modules/**/*')}`,
+  `!${path.join(processCwd, '**/+(node_modules|bower_components|vendor|dist)/**/*')}`,
   path.join(processCwd, 'bin/**/*.js'),
   path.join(processCwd, 'src/**/*.+(js|ts|tsx)'),
   path.join(processCwd, 'gulp/**/*.js'),